src/detonation-chamber/requirements.txt
# Nexus-CyberAgent Detonation Chamber - Python Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# HTTP clients
httpx==0.25.0
aiohttp==3.9.0
requests==2.31.0

# Malware analysis
pefile==2023.2.7
capstone==5.0.1
unicorn==2.0.1
keystone-engine==0.9.2
ropper==1.13.8
yara-python==4.3.1
oletools==0.60.1
pdfid==1.1.0
distorm3==3.5.2

# Volatility 3
volatility3==2.5.0

# Network analysis
scapy==2.5.0
pyshark==0.6
dpkt==1.9.8
zeek==5.0.0

# Cryptography
pycryptodome==3.19.0
cryptography==41.0.5

# Cuckoo Sandbox
cuckoo==2.0.7
sqlalchemy==2.0.23
jinja2==3.1.2
chardet==5.2.0
python-magic==0.4.27
pydeep==0.4
androguard==3.4.0a1

# Database
asyncpg==0.29.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
tabulate==0.9.0
python-dateutil==2.8.2
python-multipart==0.0.6

# Parsing
xmltodict==0.13.0
pyyaml==6.0.1
jsonschema==4.19.1

# Logging and monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
